split code into own lib
